perf(charts,search): defer chart/palette bundles, type chart data, race-safe search
- AllocationChart: make the DonutChartWrapper import type-only so the dynamic()
  import is the sole code path and recharts is genuinely deferred (#225).
- Replace the loose `ChartDataPoint { [key: string]: any }` shape with domain
  types (ChartDatum, PortfolioValuePoint, AssetAllocationSlice,
  BenchmarkComparisonPoint) and make the chart wrappers generic over them (#226).
- Lazy-load the command palette: split the dialog into CommandPaletteDialog and
  load it via next/dynamic from a tiny launcher that owns the Cmd/Ctrl+K
  shortcut, keeping it out of the initial bundle; GlobalSearch stays eager since
  it renders in the navbar (#227).
- GlobalSearch: guard async results with a monotonic request id so a slow search
  can't overwrite a newer one or repopulate after clear/navigate (#228).

Closes #225
Closes #226
Closes #227
Closes #228
